pyproject.toml
[project]
name = "api-agent"
version = "0.1.0"
description = "Universal MCP server for querying GraphQL and REST APIs using natural language"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "API Agent Contributors" }]
requires-python = ">=3.11"
dependencies = [
    "arize-otel>=0.11.0",
    "duckdb>=1.0.0",
    "fastmcp>=2.13.3",
    "httpx>=0.28.1",
    "openai-agents>=0.0.3",
    "openinference-instrumentation-openai-agents>=1.4.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
]

[project.scripts]
api-agent = "api_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api_agent"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501", "E402"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.0",
]
